Update main constructor with options
The main constructor is updated with required parameters and variable
options.

The WithOptions constructor still works as it did before.

With this change, we can have sane defaults from the default constructor
and can update it with optional options like JWE.

One change introduced here is that by default the client secret is used
as part of the PKCE encryption, and this behavior is overridable with an
option.